lqmining simulations moved to core dir.
