+++ src/test/scala/io/ergodex/core/lqmining/simple/RedeemBox.scala
@@ -1,8 +1,8 @@
-package io.ergodex.core.sim.lqmining.simple
+package io.ergodex.core.lqmining.simple
 
-import io.ergodex.core.sim.BoxRuntime.NonRunnable
-import io.ergodex.core.sim.RuntimeState.withRuntimeState
-import io.ergodex.core.sim.{AnyBox, BoxSim, RuntimeState, TryFromBox}
+import io.ergodex.core.BoxRuntime.NonRunnable
+import io.ergodex.core.RuntimeState.withRuntimeState
+import io.ergodex.core.{AnyBox, BoxSim, RuntimeState, TryFromBox}
 import io.ergodex.core.syntax._
 
 final class RedeemBox[F[_]: RuntimeState](
