+++ src/test/scala/io/ergodex/core/lqmining/simple/StakingBundleBoxSpec.scala
@@ -1,10 +1,10 @@
-package io.ergodex.core.sim.lqmining.simple
+package io.ergodex.core.lqmining.simple
 
-import io.ergodex.core.sim.Helpers.{boxId, tokenId}
-import io.ergodex.core.sim.ToLedger._
-import io.ergodex.core.sim.lqmining.simple.LMPool._
-import io.ergodex.core.sim.lqmining.simple.Token._
-import io.ergodex.core.sim.{LedgerPlatform, RuntimeCtx, SigmaProp}
+import io.ergodex.core.Helpers.{boxId, tokenId}
+import io.ergodex.core.ToLedger._
+import io.ergodex.core.lqmining.simple.LMPool._
+import io.ergodex.core.lqmining.simple.Token._
+import io.ergodex.core.{LedgerPlatform, RuntimeCtx}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
 import org.scalatestplus.scalacheck.ScalaCheckPropertyChecks
